requirements.txt
modal>=0.73.0
PyJWT>=2.8.0
supabase>=2.10.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0
pydantic-settings>=2.3.0
httpx>=0.27.0
